Fixed issues on the player and the audio actor
The player had an issue with the movement speed being too slow along with the turn speed being too intense which induced motion sickness.

The audio actor had an issue with the collisions being active when they were meant to be deactivated.  Lighting still needs adjusted to not cause migrains